fix: tms renamed to timing
