+++ src/src/cpp/logging/CMakeLists.txt
@@ -9,13 +9,13 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 # (Note that we include core/event_dispatcher.tpp as in the SCons script.)
 set(LOGGING_SOURCES
     storage/selection/selection_storage.cc
-    storage/tms/timing_storage.cc
+    storage/timing/timing_storage.cc
     loggers/selection/selection_logger.cc
-    loggers/tms/timing_logger.cc
+    loggers/timing/timing_logger.cc
     metrics/selection/selection_metrics.cc
     metrics/selection/selection_metrics_builder.cc
-    metrics/tms/timing_metrics.cc
-    metrics/tms/timing_metrics_builder.cc
+    metrics/timing/timing_metrics.cc
+    metrics/timing/timing_metrics_builder.cc
     core/event_dispatcher.tpp
 )
