+++ src/src/cpp/logging/SConscript
@@ -15,20 +15,20 @@ env.Library(
     source = [
         # Storage
         'storage/selection/selection_storage.cc',
-        'storage/tms/timing_storage.cc',
+        'storage/timing/timing_storage.cc',
         
         # Core
         'core/event_dispatcher.tpp',
         
         # Loggers
         'loggers/selection/selection_logger.cc',
-        'loggers/tms/timing_logger.cc',
+        'loggers/timing/timing_logger.cc',
         
         # Metrics
         'metrics/selection/selection_metrics.cc',
         'metrics/selection/selection_metrics_builder.cc',
-        'metrics/tms/timing_metrics.cc',
-        'metrics/tms/timing_metrics_builder.cc' 
+        'metrics/timing/timing_metrics.cc',
+        'metrics/timing/timing_metrics_builder.cc' 
     ],
     CPPPATH = ['.']
 )
