+++ src/src/cpp/logging/loggers/timing/timing_logger.h
@@ -1,6 +1,6 @@
 #include "storage/csv_storage.h"
 #include "core/event_dispatcher.h"
-#include "metrics/tms/timing_metrics.h"
+#include "metrics/timing/timing_metrics.h"
 
 
 class TimingLogger {
