+++ src/src/cpp/logging/storage/timing/timing_storage.h
@@ -1,6 +1,6 @@
 #pragma once
 #include "storage/csv_storage.h"
-#include "metrics/tms/timing_metrics.h"
+#include "metrics/timing/timing_metrics.h"
 
 
 class TimingStorage : public CSVStorage<TimingMetrics> {
